fix: general csv formatting
This update includes an additional column for the messenger's type, might be user, writer, or whatever yle comes up with in the future :)

Additionally, quotation marks have all been replaced by single quotation marks.

The chat looks a bit nicer now.

Apparently I linked `presidentinvaalit2024` as one of the example streams, when in reality it was `presidentinvaalit24`. Oops.